+++ api/helpers/songs/remove.js
@@ -24,11 +24,6 @@ module.exports = {
       type: 'boolean',
       defaultsTo: false,
       description: 'If false, system will clear the entire queue and re-queue tracks that meet criteria. If true, system will remove tracks that fail specified criteria one by one instead of clearing the entire queue and re-queuing appropriate tracks.'
-    },
-    includeCurrentTrack: {
-      type: 'boolean',
-      defaultsTo: false,
-      description: 'DEPRECATED'
     }
   },