+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Removed deprecated Songs.pendingCmd [Issue 59](https://github.com/Lovinity/wwsu/issues/59)
  - Removed resume parameter from sails.helpers.playlists.start [Issue 58](https://github.com/Lovinity/wwsu/issues/58)
  - Removed Meta.djcontrols [Issue 57](https://github.com/Lovinity/wwsu/issues/57)
+ - Removed includeCurrentTrack parameter from sails.helpers.songs.remove [Issue 56](https://github.com/Lovinity/wwsu/issues/56)
 
 ### Added
  - "device" property returned for recipient objects in recipients/get. This will contain the OneSignal UUID if the recipient is using the WWSU mobile app. Otherwise, it will be null.